requirements.txt
matplotlib~=3.8.0
numpy~=1.26.4
scikit-learn~=1.2.2
torchvision~=0.20.1
pillow~=10.2.0
torch